+++ frp-client/tests/end_to_end.rs
@@ -185,6 +185,7 @@ async fn test_e2e_tcp_proxy_over_websocket() {
             health_check_http_headers: std::collections::HashMap::new(),
             proxy_protocol_version: String::new(),
             enabled: true,
+            disable_assisted_addrs: false,
         }],
         ..Default::default()
     };
@@ -323,6 +324,7 @@ async fn test_e2e_tcp_proxy_over_tls() {
             health_check_http_headers: std::collections::HashMap::new(),
             proxy_protocol_version: String::new(),
             enabled: true,
+            disable_assisted_addrs: false,
         }],
         ..Default::default()
     };
@@ -446,6 +448,7 @@ async fn test_e2e_tcp_proxy_over_yamux() {
             health_check_http_headers: std::collections::HashMap::new(),
             proxy_protocol_version: String::new(),
             enabled: true,
+            disable_assisted_addrs: false,
         }],
         ..Default::default()
     };